docs: add measurement usage guide to index
This is adapted from the original `measurement-framework.md` page
which basically had the new text plus a copy of each measurement
table from each page. This felt a bit redundant and would be hard
to maintain so I've moved the measurement framework description
to the index and left it so that you go to each page to see the
measurement table.